+++ kernel/tests/integration/features/mod.rs
@@ -6,4 +6,5 @@ mod clustering_e2e;
 mod dv;
 mod iceberg_compat;
 mod maintenance_ops;
+mod materialize_partition_columns;
 mod row_tracking;
