+++ kernel/tests/integration/write/column_mapping.rs
@@ -17,7 +17,6 @@ use delta_kernel::object_store::{DynObjectStore, ObjectStoreExt as _};
 use delta_kernel::schema::{DataType, StructField, StructType};
 use delta_kernel::table_features::{get_any_level_column_physical_name, ColumnMappingMode};
 use delta_kernel::{Engine, FileMeta, Snapshot};
-use test_utils::table_builder::{FeatureSet, TestTableBuilder};
 use test_utils::{
     assert_partition_values, assert_schema_has_field, copy_directory,
     create_table_and_load_snapshot, nested_batches, nested_schema, read_actions_from_commit,
@@ -381,176 +380,6 @@ async fn test_column_mapping_partitioned_write(
     for remove in &remove_actions {
         assert_partition_values(remove, &physical_name, "A");
     }
-
-    Ok(())
-}
-
-// End-to-end: column-mapped + materializePartitionColumns + partition cols in the middle
-// of the logical schema. Asserts parquet field order = data cols first then partition cols,
-// and verifies checkpoint + scan round trip works as expected.
-#[rstest::rstest]
-#[case::cm_none(ColumnMappingMode::None)]
-#[case::cm_id(ColumnMappingMode::Id)]
-#[case::cm_name(ColumnMappingMode::Name)]
-#[tokio::test(flavor = "multi_thread")]
-async fn test_column_mapping_materialized_partition_in_middle_roundtrip(
-    #[case] cm_mode: ColumnMappingMode,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt::try_init();
-
-    // Logical schema: data cols interleaved with partition cols.
-    let schema = Arc::new(StructType::try_new(vec![
-        StructField::nullable("a", DataType::INTEGER),
-        StructField::nullable("p1", DataType::STRING),
-        StructField::nullable("b", DataType::INTEGER),
-        StructField::nullable("p2", DataType::STRING),
-        StructField::nullable("c", DataType::INTEGER),
-    ])?);
-
-    // Step 1: CREATE TABLE with column mapping + materializePartitionColumns via TestTableBuilder.
-    let cm_str = match cm_mode {
-        ColumnMappingMode::None => "none",
-        ColumnMappingMode::Id => "id",
-        ColumnMappingMode::Name => "name",
-    };
-    let table = TestTableBuilder::new()
-        .with_schema(schema.clone())
-        .with_partition_columns(["p2", "p1"])
-        .with_features(
-            FeatureSet::new()
-                .column_mapping(cm_str)
-                .with_property("delta.feature.materializePartitionColumns", "supported"),
-        )
-        .build()?;
-
-    let engine = Arc::new(
-        DefaultEngineBuilder::new(table.store().clone())
-            .with_task_executor(Arc::new(TokioMultiThreadExecutor::new(
-                tokio::runtime::Handle::current(),
-            )))
-            .build(),
-    );
-    let table_url = Url::parse(table.table_root())?;
-    let snapshot = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-
-    // Step 2: Write one batch carrying data with the full logical schema. Materialization keeps the
-    // partition columns in the parquet, sourced from the input batch via the L2P transform.
-    let arrow_schema = Arc::new(schema.as_ref().try_into_arrow()?);
-    let batch = RecordBatch::try_new(
-        arrow_schema,
-        vec![
-            Arc::new(Int32Array::from(vec![1, 2, 3])),
-            Arc::new(StringArray::from(vec!["x", "x", "x"])),
-            Arc::new(Int32Array::from(vec![10, 20, 30])),
-            Arc::new(StringArray::from(vec!["y", "y", "y"])),
-            Arc::new(Int32Array::from(vec![100, 200, 300])),
-        ],
-    )?;
-    let partition_values = HashMap::from([
-        ("p1".to_string(), Scalar::String("x".into())),
-        ("p2".to_string(), Scalar::String("y".into())),
-    ]);
-    let post_write_snapshot =
-        write_batch_to_table(&snapshot, engine.as_ref(), batch, partition_values).await?;
-
-    // Step 3: Validate the physical parquet layout: data cols first, partition cols last,
-    // all under physical names. Field order in `footer_schema` mirrors the parquet column
-    // order on disk.
-    let logical_schema = snapshot.schema();
-    let physical_name = |logical: &str| -> Result<String, Box<dyn std::error::Error>> {
-        let mut names = get_any_level_column_physical_name(
-            logical_schema.as_ref(),
-            &ColumnName::new([logical]),
-            cm_mode,
-        )?
-        .into_inner();
-        Ok(names.remove(0))
-    };
-    let expected_physical_order = [
-        physical_name("a")?,
-        physical_name("b")?,
-        physical_name("c")?,
-        physical_name("p2")?,
-        physical_name("p1")?,
-    ];
-
-    let add_actions = read_add_infos(&post_write_snapshot, engine.as_ref())?;
-    assert_eq!(add_actions.len(), 1, "expected one add file");
-    let parquet_url = table_url.join(&add_actions[0].path)?;
-    let obj_meta = table
-        .store()
-        .head(&Path::from_url_path(parquet_url.path())?)
-        .await?;
-    let file_meta = FileMeta::new(
-        parquet_url,
-        0, /* last_modified */
-        obj_meta.size as u64,
-    );
-    let footer = engine.parquet_handler().read_parquet_footer(&file_meta)?;
-    let footer_field_names: Vec<&str> = footer.schema.fields().map(|f| f.name.as_str()).collect();
-    assert_eq!(
-        footer_field_names,
-        expected_physical_order
-            .iter()
-            .map(|s| s.as_str())
-            .collect::<Vec<_>>(),
-        "parquet field order must be data cols first, then partition cols (in \
-         partition_columns order), under physical names"
-    );
-
-    // Step 4: Checkpoint, reload a fresh snapshot, scan, and assert the logical schema +
-    // values survive the round-trip. Logical column names and order must come back intact.
-    post_write_snapshot.checkpoint(engine.as_ref(), None)?;
-    let reloaded = Snapshot::builder_for(table_url.clone()).build(engine.as_ref())?;
-    let scan = reloaded.scan_builder().build()?;
-    let batches: Vec<RecordBatch> = scan
-        .execute(engine.clone())?
-        .map(|r| {
-            let arrow = ArrowEngineData::try_from_engine_data(r.unwrap()).unwrap();
-            arrow.record_batch().clone()
-        })
-        .collect();
-    let result_schema = batches[0].schema();
-    let combined = delta_kernel::arrow::compute::concat_batches(&result_schema, &batches)?;
-    let combined_schema = combined.schema();
-    let logical_field_names: Vec<&str> = combined_schema
-        .fields()
-        .iter()
-        .map(|f| f.name().as_str())
-        .collect();
-    assert_eq!(
-        logical_field_names,
-        vec!["a", "p1", "b", "p2", "c"],
-        "reloaded scan must return the logical schema field order"
-    );
-
-    let combined_struct = StructArray::from(combined);
-    let a: &Int32Array = resolve_struct_field(&combined_struct, &["a".into()]);
-    let p1: &StringArray = resolve_struct_field(&combined_struct, &["p1".into()]);
-    let b: &Int32Array = resolve_struct_field(&combined_struct, &["b".into()]);
-    let p2: &StringArray = resolve_struct_field(&combined_struct, &["p2".into()]);
-    let c: &Int32Array = resolve_struct_field(&combined_struct, &["c".into()]);
-
-    let mut rows: Vec<(i32, String, i32, String, i32)> = (0..a.len())
-        .map(|i| {
-            (
-                a.value(i),
-                p1.value(i).to_string(),
-                b.value(i),
-                p2.value(i).to_string(),
-                c.value(i),
-            )
-        })
-        .collect();
-    rows.sort_by_key(|r| r.0);
-    assert_eq!(
-        rows,
-        vec![
-            (1, "x".to_string(), 10, "y".to_string(), 100),
-            (2, "x".to_string(), 20, "y".to_string(), 200),
-            (3, "x".to_string(), 30, "y".to_string(), 300),
-        ],
-    );
 
     Ok(())
 }
